Make 'rake db:seed' safe to call multiple times. The 'rake db:seed' command will execute db/seeds.rb which will populate the database with startup entries. If you run the command more then once it will insert the seed entries multiple times. To make the command safe to call more then once this changeset adds a check so that it only adds the entries if the table the entries are being added to is currently empty.
In this case we are pre-populating the database with target languges. If we called 'rake db:seed' multiple
times we'd get the same langauges added to the database more then once. With this change if the
target langages table already has entries we forgo popluating it assuming that we've already seeded
the database with target languages.